move the tab resolver beside the segments it reserves
resolveTab is not a hook, so it sits with the route vocabulary in
constants/routes instead of leaving routes pointing at customHooks. The
reserved-segment check is now a constraint on the tabs it takes, so tsc
rejects a slug the record routes spend rather than a render throwing.
