+++ app/src/gui/components/notebook/plans/ListOfRecordsPlanView.tsx
@@ -14,7 +14,7 @@ import {
 } from '@mui/material';
 import {useCallback, useMemo} from 'react';
 import {RecordsTable} from '../record_table';
-import {resolveTab} from '../../../../utils/customHooks';
+import {resolveTab} from '../../../../constants/routes';
 import {NotebookViewComponentProps} from '../types';
 import {config} from '../../../../buildconfig';
